requirements.txt
numpy>=1.19.0
matplotlib>=3.3.0
Pillow>=8.0.0
tifffile>=2020.9.3
PyOpenGL>=3.1.5
PyOpenGL-accelerate>=3.1.5
pyopengltk>=0.0.4
tkinter-tooltip>=2.0.0 